services/camera-ingest/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pyzmq==25.1.1
redis==5.0.1
PyYAML==6.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
